Revert "Move `FTPSUploader` to `core.uploader.base` and make private"
This reverts commit 6ce589e21a01eda9075c8233473c669ca361c5cc.
